+++ vue-components/src/components/Button.vue
@@ -3,8 +3,8 @@
 		:class="[
 			'wikit',
 			'wikit-Button',
-			`wikit-Button--${ type }`,
-			`wikit-Button--${ variant }`,
+			`wikit-Button--${type}`,
+			`wikit-Button--${variant}`,
 			iconOnly ? `wikit-Button--iconOnly` : '',
 		]"
 	>
@@ -14,22 +14,23 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import VueCompositionAPI, { defineComponent, onMounted } from '@vue/composition-api';
+import VueCompositionAPI, {
+	defineComponent,
+	onMounted,
+} from '@vue/composition-api';
 
 Vue.use( VueCompositionAPI );
 
-/* eslint-disable no-trailing-spaces */
 /**
  *  An interactive element signaling a single-step action that will occur when the user clicks or taps on it.
- * 
+ *
  * Known issues:
- * 
- * * The styles defined on `:focus` do not apply in Safari and Firefox on macOS.  
- * This seems to be a desired behavior and not a bug  
- * https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#Clicking_and_focus  
+ *
+ * * The styles defined on `:focus` do not apply in Safari and Firefox on macOS.<br>
+ * This seems to be a desired behavior and not a bug<br>
+ * https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button#Clicking_and_focus<br>
  * https://bugzilla.mozilla.org/show_bug.cgi?id=1581369#c5
  */
-/* eslint-enable no-trailing-spaces */
 export default defineComponent( {
 	name: 'Button',
 	props: {
@@ -48,12 +49,12 @@ export default defineComponent( {
 		/**
 		 * The variant of the button
 		 *
-		 * Allowed values: `normal`, `primary`
+		 * Allowed values: `normal`, `primary`, `quiet`
 		 */
 		variant: {
 			type: String,
 			validator( value: string ): boolean {
-				return [ 'normal', 'primary' ].includes( value );
+				return [ 'normal', 'primary', 'quiet' ].includes( value );
 			},
 			default: 'normal',
 		},
@@ -63,13 +64,14 @@ export default defineComponent( {
 		},
 	},
 	setup( props: {
-		type: 'neutral'|'progressive'|'destructive';
-		variant: 'normal'|'primary';
+		type: 'neutral' | 'progressive' | 'destructive';
+		variant: 'normal' | 'primary' | 'quiet';
 	} ) {
 		onMounted( () => {
 			const supportedCombinations = {
 				normal: [ 'neutral' ],
 				primary: [ 'progressive', 'destructive' ],
+				quiet: [ 'neutral', 'progressive', 'destructive' ],
 			};
 			if ( !supportedCombinations[ props.variant ].includes( props.type ) ) {
 				throw new Error(
@@ -186,6 +188,91 @@ $base: '.wikit-Button';
 		color: $wikit-Button-normal-disabled-color;
 		background-color: $wikit-Button-normal-disabled-background-color;
 		border-color: $wikit-Button-normal-disabled-border-color;
+		cursor: default;
+		pointer-events: none;
+	}
+
+	&#{$base}--quiet {
+		&#{$base}--neutral {
+			color: $wikit-Button-quiet-neutral-color;
+			background-color: $wikit-Button-quiet-background-color;
+			border-color: $wikit-Button-quiet-border-color;
+
+			&:hover {
+				color: $wikit-Button-quiet-neutral-hover-color;
+				background-color: $wikit-Button-quiet-neutral-hover-background-color;
+				border-color: $wikit-Button-quiet-neutral-hover-border-color;
+			}
+
+			&:active {
+				color: $wikit-Button-quiet-neutral-active-color;
+				background-color: $wikit-Button-quiet-neutral-active-background-color;
+				border-color: $wikit-Button-quiet-neutral-active-border-color;
+			}
+
+			// A clicked button is both :active and :focused. Using :not(:active) to avoid mixing the two.
+			&:focus:not(:active) {
+				color: $wikit-Button-quiet-neutral-focus-color;
+				background-color: $wikit-Button-quiet-neutral-focus-background-color;
+				border-color: $wikit-Button-quiet-neutral-focus-border-color;
+				box-shadow: $wikit-Button-quiet-neutral-focus-box-shadow;
+			}
+		}
+		&#{$base}--progressive {
+			color: $wikit-Button-quiet-progressive-color;
+			background-color: $wikit-Button-quiet-background-color;
+			border-color: $wikit-Button-quiet-border-color;
+
+			&:hover {
+				color: $wikit-Button-quiet-progressive-hover-color;
+				background-color: $wikit-Button-quiet-progressive-hover-background-color;
+				border-color: $wikit-Button-quiet-progressive-hover-border-color;
+			}
+
+			&:active {
+				color: $wikit-Button-quiet-progressive-active-color;
+				background-color: $wikit-Button-quiet-progressive-active-background-color;
+				border-color: $wikit-Button-quiet-progressive-active-border-color;
+			}
+
+			// A clicked button is both :active and :focused. Using :not(:active) to avoid mixing the two.
+			&:focus:not(:active) {
+				color: $wikit-Button-quiet-progressive-focus-color;
+				background-color: $wikit-Button-quiet-progressive-focus-background-color;
+				border-color: $wikit-Button-quiet-progressive-focus-border-color;
+				box-shadow: $wikit-Button-quiet-progressive-focus-box-shadow;
+			}
+		}
+		&#{$base}--destructive {
+			color: $wikit-Button-quiet-destructive-color;
+			background-color: $wikit-Button-quiet-background-color;
+			border-color: $wikit-Button-quiet-border-color;
+
+			&:hover {
+				color: $wikit-Button-quiet-destructive-hover-color;
+				background-color: $wikit-Button-quiet-destructive-hover-background-color;
+				border-color: $wikit-Button-quiet-destructive-hover-border-color;
+			}
+
+			&:active {
+				color: $wikit-Button-quiet-destructive-active-color;
+				background-color: $wikit-Button-quiet-destructive-active-background-color;
+				border-color: $wikit-Button-quiet-destructive-active-border-color;
+			}
+
+			// A clicked button is both :active and :focused. Using :not(:active) to avoid mixing the two.
+			&:focus:not(:active) {
+				color: $wikit-Button-quiet-destructive-focus-color;
+				background-color: $wikit-Button-quiet-destructive-focus-background-color;
+				border-color: $wikit-Button-quiet-destructive-focus-border-color;
+				box-shadow: $wikit-Button-quiet-destructive-focus-box-shadow;
+			}
+		}
+	}
+	&#{$base}--quiet:disabled {
+		color: $wikit-Button-quiet-disabled-color;
+		background-color: $wikit-Button-quiet-background-color;
+		border-color: $wikit-Button-quiet-border-color;
 		cursor: default;
 		pointer-events: none;
 	}